rename: api 모듈로 분리 #59
