+++ src/app/api/reserve/fetchReserveInfo.ts
@@ -10,7 +10,7 @@ export const fetchReserveInfo = async (reservationId: string) => {
     .single();
 
   if (error) {
-    console.error('예약 정보 불러오기 오류 => ', error);
+    console.error('예약 정보 불러오기 오류 발생 => ', error);
     return;
   }
